+++ README.md
@@ -25,6 +25,6 @@ $ java -jar amazon-kinesis-replay.jar -streamArn «Kinesis stream ARN» -bucketN
 
 More options can be obtained through the `-help` parameter.
 
-## License Summ
+## License Summary
 
 This sample code is made available under a modified MIT license. See the LICENSE file.
