+++ pom.xml
@@ -125,7 +125,7 @@
                             <transformers>
                                 <transformer
                                         implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>com.amazonaws.samples.kinesis.replay2.RePlayer</mainClass>
+                                    <mainClass>com.amazonaws.samples.kinesis.replay.RePlayer</mainClass>
                                 </transformer>
                             </transformers>
                         </configuration>
