+++ src/main/java/com/amazonaws/samples/kinesis/replay/events/JsonEvent.java
@@ -15,7 +15,7 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.amazonaws.samples.kinesis.replay2.events;
+package com.amazonaws.samples.kinesis.replay.events;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
