+++ src/main/java/com/amazonaws/samples/kinesis/replay/utils/JsonEventBufferedReader.java
@@ -15,14 +15,13 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.amazonaws.samples.kinesis.replay2.utils;
+package com.amazonaws.samples.kinesis.replay.utils;
 
-import com.amazonaws.samples.kinesis.replay2.events.JsonEvent;
+import com.amazonaws.samples.kinesis.replay.events.JsonEvent;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.UUID;
-import java.util.concurrent.ExecutionException;
 import java.util.concurrent.PriorityBlockingQueue;
 import java.util.concurrent.Semaphore;
