+++ src/main/resources/log4j2.properties
@@ -23,11 +23,11 @@ appender.console.type = CONSOLE
 appender.console.layout.type = PatternLayout
 appender.console.layout.pattern = %d{HH:mm:ss,SSS} %-5p %-32c{1.} %x - %m%n
 
-#logger.replay2.name=com.amazonaws.samples.kinesis.replay2
-#logger.replay2.level=DEBUG
+#logger.replay.name=com.amazonaws.samples.kinesis.replay
+#logger.replay.level=DEBUG
 
-#logger.producer.name=com.amazonaws.samples.kinesis.replay2.utils.KinesisProducer
+#logger.producer.name=com.amazonaws.samples.kinesis.replay.utils.KinesisProducer
 #logger.producer.level=DEBUG
 
-#logger.replayer.name=com.amazonaws.samples.kinesis.replay2.RePlayer
+#logger.replayer.name=com.amazonaws.samples.kinesis.replay.RePlayer
 #logger.replayer.level=DEBUG
